server organized into modules
