+++ server/colorsource.js
@@ -22,8 +22,8 @@ const sendColor = ()=> {
     // console.log('in sendColor')
     axios({
             method: 'POST',
-            url: 'https://sci-fair-scum.herokuapp.com/color',
-            // url: 'http://localhost:5000/color',
+            // url: 'https://sci-fair-scum.herokuapp.com/color',
+            url: 'http://localhost:5000/color',
             data: newColor()
         })
     .then((res)=>{
